+++ source/renderer/app/components/voting/voting-info/RegisterToVote.tsx
@@ -1,18 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
+import moment from 'moment';
 import { injectIntl } from 'react-intl';
-import { Button } from 'react-polymorph/lib/components/Button';
-import { Checkbox } from 'react-polymorph/lib/components/Checkbox';
-import {
-  formattedDateTime,
-  mapToLongDateTimeFormat,
-} from '../../../utils/formatters';
-import type { Locale } from '../../../../../common/types/locales.types';
 import type { Intl } from '../../../types/i18nTypes';
+
+import type { Locale } from '../../../../../common/types/locales.types';
 import { messages } from './RegisterToVote.messages';
 import { messages as votingMessages } from './VotingInfo.messages';
 import styles from './RegisterToVote.scss';
-import votingStyles from './VotingInfo.scss';
 import type { CatalystFund } from '../../../api/voting/types';
+import { logger } from '../../../utils/logging';
 
 type Props = {
   currentLocale: Locale;
@@ -23,7 +19,20 @@ type Props = {
   onRegisterToVoteClick: (...args: Array<any>) => any;
 };
 
-function RegisterToVote({
+const isNextFundToBeDefined = (registrationSnapshotTime: Date): boolean => {
+  try {
+    return moment().diff(registrationSnapshotTime) > 0;
+  } catch (error) {
+    logger.error('Voting::NextFund::Invalid date', {
+      error,
+    });
+  }
+  return false;
+};
+
+const NextFund = React.lazy(() => import('./NextFund'));
+
+export function RegisterToVote({
   currentLocale,
   currentDateFormat,
   currentTimeFormat,
@@ -31,20 +40,10 @@ function RegisterToVote({
   intl,
   onRegisterToVoteClick,
 }: Props) {
-  const [step1, setStep1] = useState(false);
-  const [step2, setStep2] = useState(false);
-  const canRegister = step1 && step2;
-  const snapshotDate = formattedDateTime(
-    fundInfo.next.registrationSnapshotTime,
-    {
-      currentLocale,
-      ...mapToLongDateTimeFormat({
-        currentLocale,
-        currentDateFormat,
-        currentTimeFormat,
-      }),
-    }
+  const isNextFundDefined = isNextFundToBeDefined(
+    fundInfo?.next?.registrationSnapshotTime
   );
+  // const isNextFundDefined = false;
   return (
     <div className={styles.root}>
       <span className={styles.title}>
@@ -52,36 +51,23 @@ function RegisterToVote({
           votingFundNumber: fundInfo.next.number,
         })}
       </span>
-      <span className={styles.dateLabel}>
-        {intl.formatMessage(messages.dateLabel)}
-      </span>
-      <span className={styles.date}>{snapshotDate}</span>
-      <hr className={votingStyles.separator} />
-      <span className={styles.stepsTitle}>
-        {intl.formatMessage(messages.stepsTitle)}
-      </span>
-      <div className={styles.step}>
-        <Checkbox
-          checked={step1}
-          onChange={setStep1}
-          label={intl.formatMessage(messages.step1CheckBoxLabel)}
+      {isNextFundDefined ? (
+        <NextFund
+          currentLocale={currentLocale}
+          currentDateFormat={currentDateFormat}
+          currentTimeFormat={currentTimeFormat}
+          fundInfo={fundInfo}
+          onRegisterToVoteClick={onRegisterToVoteClick}
         />
-      </div>
-      <div className={styles.step}>
-        <Checkbox
-          checked={step2}
-          label={intl.formatMessage(messages.step2CheckBoxLabel)}
-          onChange={setStep2}
-        />
-      </div>
-      <Button
-        className={styles.button}
-        onClick={onRegisterToVoteClick}
-        label={intl.formatMessage(messages.buttonLabel)}
-        disabled={!canRegister}
-      />
+      ) : (
+        <React.Fragment>
+          <span className={styles.dateLabel}>
+            {intl.formatMessage(messages.dateLabel)}
+          </span>
+          <span className={styles.date}>{'To be defined'}</span>
+        </React.Fragment>
+      )}
     </div>
   );
 }
-
 export default injectIntl(RegisterToVote);