+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Chores
 
+- Showing 'To be defined' if date of the next Catalyst fund is in the past ([PR 3105](https://github.com/input-output-hk/daedalus/pull/3105))
 - Add members of Cardano Wallet team to the "About Daedalus" screen ([PR 3092](https://github.com/input-output-hk/daedalus/pull/3092))
 - Removed Chromatic from project ([PR 3126](https://github.com/input-output-hk/daedalus/pull/3126))
 - Upgraded Electron to 24.2, and Node.js to 18.× ([PR 3124](https://github.com/input-output-hk/daedalus/pull/3124))